Fix: snapshot refactoring (#3022)
* remove iPhone 4 and add iPhone X*

* move constant to header

* change default iPhone size to iPhone 4 inch

* translate methods in Swift to prevent dependency issue

* rename methods

* add deviceName argument

* pass device name

* clean up comment

* regenerate snapshots
remove iPhone 4 snapshots
add iPhone 2018 snapshots

* clean up

* clean up

* undo snapshot with emoji

* remove not in use interface

* clean up

* clean up header

* clean up

* translate verify function to Swift

* remove the Purelayout

* remove Cartography (10 second faster?)

* rename

* remove cartography
update AudioRecordViewControllerTests ref image for different drawing resolution produced

* update ConversationInputBarViewControllerTests

* update snapshots

* clean up

* expend FBSnapshotVerifyViewWithOptions macro

* remove cartography

* rename snapshot file name

* rename files

* convert snapshotVerifyViewOrLayerWithOptions to Swift.

The failed snapshot test has info for file and line

* translate assertAmbigousLayout to Swift

* update snapshots

* tolerance in AudioRecordViewControllerTests

* update snapshots

* remove optional in error message

* fix snapshots

* translate setup method

* Swiftify remain methods

* Convert ZMSnapshotTestCase to Swift

* clean up
fix accent color

* clean up

* import XCTest after @c header is deleted

* fix array compare method

* convert constants to Swift
update snapshots

* patch ConversationListBottomBarControllerTests for its zero height issue

* remove unnecessary layout

* scroll textview to bottom to make sure the snapshot is constant

* rename file temporally for review comparsion

* WIP fixing AdaptiveFormViewController but crash with UIViewControllerHierarchyInconsistency

* clean up

* update snapshot after removing unnecessary layout call

* WIP broken constraint in input bar

* clean up

* add config block argument to verifyInAllPhoneWidths

* rename

* remove SUT init twice issue

* regenerate snapshots

* tolerance in ProfileClientViewControllerTests

* scroll textview to bottom after view resizes

* set view.layer.speed = 0 for all tests

* fix invalid touch change y offset

* update snapshots

* test finishes in 1m 55s

* tolerance i testThatItRendersMentionWithEmoji_MultipleMention

* extraLayoutPass for testThatItShowsUserWithImage

* WIP TextView constraints

* fix placeholder constraints

* WIP fix the TextView constraint

* text container inset

* fix placeholder constraints
fix verifyInIPhoneSize()

* confirm that InputBar does not need to custom placeholderTextContainerInset

* update snapshots

* WIP scroll to bottom

* do not scroll to bottom

* clean up

* clean up

* clean up

* clean up

* do not create placeholder label when draw rect

* comment

* WIP config after resize view

* refactoring

* update snapshots

* update snapshot for long text

* fix merge issues

* update snapshots

* WIP testing cell color in dark mode

* fix background color

* rewrite verifyInAllPhoneWidths in Swift
refactor ConversationInputBarViewController_AudioRecorderSnapshotTests for correct result in real running cases

* clean up

* clean up

* clean up

* clean up

* update snapshot view with higher sampling waveform

* tolerance: 0.01

* adjust tolerance

* fix merge

* update snapshots

* update snapshot

* fix waitForGroupsToBeEmpty may fail issue

* translate ZMSnapshotTestCase to Swift

* clean up

* refactor XCTAssertArrayEqual argument type to fix compile error

* fix flaky test

* rename

* clean up duplicated methods

* remove duplicated methods

* comment
