+++ Wire-iOS Tests/CallHapticsControllerTests.swift
@@ -16,7 +16,7 @@
 // along with this program. If not, see http://www.gnu.org/licenses/.
 //
 
-import Foundation
+import XCTest
 @testable import Wire
 
 fileprivate class MockCallHapticsGenerator: CallHapticsGeneratorType {
