+++ Wire-iOS Tests/CanvasViewControllerTests.swift
@@ -37,7 +37,6 @@ final class CanvasViewControllerTests: ZMSnapshotTestCase {
 
     func testForSendButtonEnalbed(){
         sut.sendButton.isEnabled = true
-        sut.view.prepareForSnapshot()
         verify(view: sut.view)
     }
