+++ Wire-iOS Tests/ConversationListBottomBarControllerTests.swift
@@ -52,6 +52,11 @@ final class ConversationListBottomBarControllerTests: ZMSnapshotTestCase {
         mockDelegate = MockConversationListBottomBarDelegate()
         UIView.performWithoutAnimation({
             self.sut = ConversationListBottomBarController(delegate: self.mockDelegate)
+
+            ///SUT has a priority 750 height constraint. fix its height first
+            NSLayoutConstraint.activate([
+                sut.view.heightAnchor.constraint(equalToConstant: 56)
+                ])
         })
     }
