+++ Wire-iOS Tests/ConversationListCellTests.swift
@@ -18,6 +18,7 @@
 
 import UIKit
 @testable import Wire
+import XCTest
 
 class ConversationListCellTests: CoreDataSnapshotTestCase {
 
@@ -47,7 +48,6 @@ class ConversationListCellTests: CoreDataSnapshotTestCase {
         ) {
         sut.conversation = conversation
         
-        sut.prepareForSnapshot()
         verify(view: sut, file: file, line: line)
     }
     
@@ -296,7 +296,7 @@ class ConversationListCellTests: CoreDataSnapshotTestCase {
         guard let conversation = conversation else { XCTFail(); return }
         sut.conversation = conversation
         sut.itemView.rightAccessory.icon = icon
-        sut.prepareForSnapshot()
+
         verify(view: sut)
     }
