+++ Wire-iOS Tests/InputBar/InputBarTests.swift
@@ -46,11 +46,10 @@ class InputBarTests: ZMSnapshotTestCase {
 
     override func setUp() {
         super.setUp()
+
         sut = InputBar(buttons: buttons())
         sut.leftAccessoryView.isHidden = true
         sut.rightAccessoryStackView.isHidden = true
-        sut.translatesAutoresizingMaskIntoConstraints = false
-        sut.layer.speed = 0
     }
 
     override func tearDown() {
@@ -77,8 +76,7 @@ class InputBarTests: ZMSnapshotTestCase {
     
     func testLongText() {
         sut.textView.text = longText
-        
-        
+
         verifyInAllPhoneWidths(view: sut)
         verifyInAllTabletWidths(view: sut)
     }
@@ -116,9 +114,7 @@ class InputBarTests: ZMSnapshotTestCase {
         inputBar.rightAccessoryStackView.isHidden = true
 
         inputBar.translatesAutoresizingMaskIntoConstraints = false
-        inputBar.layer.speed = 0
-        
-        
+
         verifyInAllPhoneWidths(view: inputBar)
     }
     
@@ -128,9 +124,7 @@ class InputBarTests: ZMSnapshotTestCase {
         inputBar.leftAccessoryView.isHidden = true
         inputBar.rightAccessoryStackView.isHidden = true
         inputBar.textView.text = ""
-        inputBar.layer.speed = 0
-        
-        
+
         verifyInAllPhoneWidths(view: inputBar)
     }
     
@@ -160,8 +154,9 @@ class InputBarTests: ZMSnapshotTestCase {
     
     func testThatItRendersCorrectlyInEditState_LongText() {
         sut.setInputBarState(.editing(originalText: longText, mentions: []), animated: false)
-        
+
+        sut.textView.resignFirstResponder() // make sure to avoid cursor being visible
+
         verifyInAllPhoneWidths(view: sut)
     }
-
 }
