+++ Wire-iOS Tests/TransformTextViewTests.swift
@@ -16,7 +16,7 @@
 // along with this program. If not, see http://www.gnu.org/licenses/.
 //
 
-import UIKit
+import XCTest
 
 class TransformTextViewTests: XCTestCase {
