+++ Wire-iOS Tests/Utils/ZMSnapshotTestCase+Internal.h
@@ -16,21 +16,6 @@
 // along with this program. If not, see http://www.gnu.org/licenses/.
 //
 
-static CGSize const ZMDeviceSizeIPhone5 = (CGSize){ .width = 320, .height = 568 };
-static CGSize const ZMDeviceSizeIPhone6 = (CGSize){ .width = 375, .height = 667 };
-static CGSize const ZMDeviceSizeIPhone6Plus = (CGSize){ .width = 414, .height = 736 };
-static CGSize const ZMDeviceSizeIPhoneX = (CGSize){ .width = 375, .height = 812 };
-static CGSize const ZMDeviceSizeIPhoneXR = (CGSize){ .width = 414, .height = 896 };
-
-static CGSize const ZMDeviceSizeIPadPortrait = (CGSize){ .width = 768, .height = 1024 };
-static CGSize const ZMDeviceSizeIPadLandscape = (CGSize){ .width = 1024, .height = 768 };
-
-static NSArray<NSValue *> * _Nonnull tabletSizes(void) {
-    return @[
-             [NSValue valueWithCGSize:ZMDeviceSizeIPadPortrait],
-             [NSValue valueWithCGSize:ZMDeviceSizeIPadLandscape]
-             ];
-}
 
 // expose _autolayoutTrace for Swift
 @interface UIView (LayoutDebugging)
@@ -40,3 +25,9 @@ static NSArray<NSValue *> * _Nonnull tabletSizes(void) {
 - (_Nullable id)_autolayoutTrace;
 
 @end
+
+// expose PersonName.stringsToPersonNames for Swift
+@interface PersonName (stringsToPersonNames)
++ (NSCache<NSString *, PersonName *> * _Nonnull)stringsToPersonNames SWIFT_WARN_UNUSED_RESULT;
+@end
+
