+++ Wire-iOS Tests/Wire-iOS-Tests-Bridging-Header.h
@@ -25,8 +25,7 @@
 #import "Analytics.h"
 #import "ParticipantDeviceCell.h"
 #import "NSString+Fingerprint.h"
-#import "ZMSnapshotTestCase.h"
-#import "XCTestCase+Internal.h"
+#import "ZMSnapshotTestCase+Internal.h"
 #import "MockUser.h"
 #import "SplitViewController.h"
 #import <avs/AVSAudioEffect.h>
