+++ Wire-iOS/Sources/UserInterface/Helpers/UIColor+WAZExtensions.h
@@ -26,6 +26,7 @@
 
 /// Set override accent color. Can set to ZMAccentColorUndefined to remove override.
 + (void)setAccentOverrideColor:(ZMAccentColor)overrideColor;
++ (ZMAccentColor)accentOverrideColor;
 
 /// Set accent color on self user to this index.
 + (void)setAccentColor:(ZMAccentColor)accentColor;
